split local and peer apis
