+++ apps/pollbook/backend/schema.sql
@@ -1,6 +1,17 @@
 CREATE TABLE voters (
     voter_id TEXT PRIMARY KEY,
+    original_first_name TEXT,
+    original_last_name TEXT,
+    updated_first_name TEXT,
+    updated_last_name TEXT,
     voter_data TEXT not null
+);
+
+CREATE TABLE machines (
+    machine_id TEXT PRIMARY KEY,
+    status TEXT not null,
+    last_updated INTEGER NOT NULL, -- last time the machine was updated
+    last_seen INTEGER NOT NULL -- last time the machine was seen
 );
 
 CREATE TABLE elections (
@@ -24,3 +35,17 @@ CREATE TABLE event_log (
 
 -- Index for sorting events by hybrid logical clock physical time then logical counter - machine id is included in the rare event of tie
 CREATE INDEX idx_hlc ON event_log (physical_time, logical_counter, machine_id); 
+CREATE INDEX idx_machine_hlc ON event_log (machine_id, physical_time, logical_counter); 
+CREATE INDEX idx_voter_hlc ON event_log (voter_id, physical_time, logical_counter, machine_id); 
+CREATE INDEX idx_updated_first_name ON voters (updated_first_name);
+CREATE INDEX idx_updated_last_name ON voters (updated_last_name);
+
+CREATE TABLE check_in_status (
+    voter_id TEXT NOT NULL,
+    machine_id TEXT,
+    is_checked_in BOOLEAN NOT NULL,
+    PRIMARY KEY (voter_id)
+);
+
+CREATE INDEX idx_check_in_status ON check_in_status (is_checked_in);
+CREATE INDEX idx_machine_check_in_status ON check_in_status (machine_id, is_checked_in);
