+++ apps/pollbook/backend/src/env.d.ts
@@ -3,7 +3,8 @@ declare namespace NodeJS {
   export interface ProcessEnv {
     readonly CI?: string;
     readonly NODE_ENV: 'development' | 'production' | 'test';
-    readonly PORT?: string;
+    readonly LOCAL_PORT?: string;
+    readonly PEER_PORT?: string;
     readonly WORKSPACE?: string;
     readonly VX_MACHINE_ID?: string;
     readonly VX_CODE_VERSION?: string;
