+++ apps/pollbook/backend/src/peer_app.ts
@@ -1,0 +1,47 @@
+import * as grout from '@votingworks/grout';
+import express, { Application } from 'express';
+import { MachineInformation, PollbookEvent, PeerAppContext } from './types';
+import {
+  fetchEventsFromConnectedPollbooks,
+  setupMachineNetworking,
+} from './networking';
+
+function buildApi(context: PeerAppContext) {
+  const { workspace, machineId } = context;
+  const { store } = workspace;
+
+  return grout.createApi({
+    getMachineInformation(): MachineInformation {
+      const election = store.getElection();
+      return {
+        machineId,
+        configuredElectionId: election ? election.id : undefined,
+      };
+    },
+
+    // TODO-CARO this is not currently called
+    receiveEvent(input: { pollbookEvent: PollbookEvent }): boolean {
+      return store.saveEvent(input.pollbookEvent);
+    },
+
+    getEvents(input: { lastEventSyncedPerNode: Record<string, number> }): {
+      events: PollbookEvent[];
+      hasMore: boolean;
+    } {
+      return store.getNewEvents(input.lastEventSyncedPerNode);
+    },
+  });
+}
+
+export type PeerApi = ReturnType<typeof buildApi>;
+
+export function buildPeerApp(context: PeerAppContext): Application {
+  const app: Application = express();
+  const api = buildApi(context);
+  app.use('/api', grout.buildRouter(api, express));
+
+  void setupMachineNetworking(context);
+  fetchEventsFromConnectedPollbooks(context);
+
+  return app;
+}
