+++ apps/pollbook/backend/src/pollbook_package.ts
@@ -13,7 +13,7 @@ import {
 import { join } from 'node:path';
 import { rootDebug } from './debug';
 import {
-  AppContext,
+  LocalAppContext,
   PollbookPackage,
   Voter,
   ValidStreetInfo,
@@ -161,7 +161,7 @@ export function pollUsbDriveForPollbookPackage({
   auth,
   workspace,
   usbDrive,
-}: AppContext): void {
+}: LocalAppContext): void {
   usbDebug('Polling USB drive for pollbook package');
   if (workspace.store.getElection()) {
     return;
