+++ apps/pollbook/backend/src/server.ts
@@ -2,28 +2,30 @@ import { useDevDockRouter } from '@votingworks/dev-dock-backend';
 import express from 'express';
 import { CITIZEN_THERMAL_PRINTER_CONFIG } from '@votingworks/printing';
 import { BaseLogger, Logger, LogSource } from '@votingworks/logging';
-import { buildApp } from './app';
-import { PORT } from './globals';
-import { AppContext } from './types';
+import { buildLocalApp } from './app';
+import { LOCAL_PORT } from './globals';
+import { LocalAppContext } from './types';
 import { getUserRole } from './auth';
 
 /**
  * Starts the server.
  */
-export function start(context: AppContext): void {
+export function start(context: LocalAppContext): void {
   const baseLogger = new BaseLogger(LogSource.VxPollbookBackend);
   const logger = Logger.from(baseLogger, () =>
     getUserRole(context.auth, context.workspace)
   );
 
-  const app = buildApp({ context, logger });
+  const app = buildLocalApp({ context, logger });
 
   useDevDockRouter(app, express, {
     printerConfig: CITIZEN_THERMAL_PRINTER_CONFIG,
   });
 
-  app.listen(PORT, () => {
+  app.listen(LOCAL_PORT, () => {
     // eslint-disable-next-line no-console
-    console.log(`VxPollbook backend running at http://localhost:${PORT}/`);
+    console.log(
+      `VxPollbook backend running at http://localhost:${LOCAL_PORT}/`
+    );
   });
 }
