+++ apps/pollbook/backend/src/voter_helpers.ts
@@ -1,0 +1,15 @@
+import { Voter } from './types';
+
+export function getUpdatedVoterFirstName(voter: Voter): string {
+  if (voter.nameChange) {
+    return voter.nameChange.firstName;
+  }
+  return voter.firstName;
+}
+
+export function getUpdatedVoterLastName(voter: Voter): string {
+  if (voter.nameChange) {
+    return voter.nameChange.lastName;
+  }
+  return voter.lastName;
+}
