+++ apps/pollbook/backend/src/workspace.ts
@@ -2,21 +2,35 @@ import { ensureDirSync } from 'fs-extra';
 import { join } from 'node:path';
 
 import { BaseLogger } from '@votingworks/logging';
-import { Workspace } from './types';
-import { Store } from './store';
+import { PeerWorkspace, LocalWorkspace } from './types';
+import { LocalStore } from './local_store';
+import { PeerStore } from './peer_store';
 
-export function createWorkspace(
+export function createLocalWorkspace(
   workspacePath: string,
   logger: BaseLogger,
   machineId: string
-): Workspace {
+): LocalWorkspace {
   ensureDirSync(workspacePath);
 
   const assetDirectoryPath = join(workspacePath, 'assets');
   ensureDirSync(assetDirectoryPath);
 
   const dbPath = join(workspacePath, 'pollbook-backend.db');
-  const store = Store.fileStore(dbPath, logger, machineId);
+  const store = LocalStore.fileStore(dbPath, logger, machineId);
 
   return { assetDirectoryPath, store };
+}
+
+export function createPeerWorkspace(
+  workspacePath: string,
+  logger: BaseLogger,
+  machineId: string
+): PeerWorkspace {
+  ensureDirSync(workspacePath);
+
+  const dbPath = join(workspacePath, 'pollbook-backend.db');
+  const store = PeerStore.fileStore(dbPath, logger, machineId);
+
+  return { store };
 }
